requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pypdf
python-markdown
python-frontmatter
ollama
faiss-cpu
typer
requests
openai
sqlalchemy
mysqlclient
pymysql
alembic
jwt
python-jose
python-dotenv
pydantic[email]
pydantic-settings
passlib
bcrypt